refactor(api): split services/leagues into a domain folder; shared isUniqueViolation (feedback #7, #10)
- packages/db/src/errors.ts owns the DrizzleQueryError -> pg DatabaseError
  unwrapping; users.ts and the join path consume it
- services/leagues.ts (668 lines, five responsibilities) becomes
  services/leagues/{locks,start,authz,join,crud,serialize,index} — verbatim
  moves, acyclic imports, barrel keeps every external import path unchanged
  and the public surface identical
