+++ apps/api/src/services/leagues/index.ts
@@ -1,0 +1,10 @@
+// Barrel preserving the `./leagues` (and `../services/leagues`) import path
+// after the service was split into a domain folder. Re-exports only the public
+// surface the old single-file service exposed — `serialize.ts` is deliberately
+// omitted: its `loadMembers`/`serializeLeague`/`LeagueRow` are module-public
+// only so `crud`/`start` can cross-import, not part of the service API.
+export * from "./locks";
+export * from "./start";
+export * from "./authz";
+export * from "./join";
+export * from "./crud";
