+++ packages/db/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./client";
+export * from "./errors";
 export * from "./schema/index";
 export * from "./queries/app-state";
